Refactot chat and transaction api and ui
