+++ client/app/agent/chat/[id]/page.tsx
@@ -70,7 +70,7 @@ export default function ChatPage() {
 
   const createNewChat = async () => {
     const id = uuidv4();
-    await router.push(`/agent/chat/${id}`);
+    await router.push(`/agent/transaction/${id}`);
   };
 
   const createNewTxn = async () => {
